chore: update imports to match next.js convention
