+++ docs/pages/docs/usage/refresh-credentials.mdx
@@ -7,7 +7,7 @@
 Use `refreshCredentials` from `next-firebase-auth-edge/lib/next/cookies` to refresh credentials in [Next.js Middleware](https://nextjs.org/docs/app/building-your-application/routing/middleware) after updating user token
 
 ```tsx filename="middleware.ts"
-import {NextRequest} from 'next/server';
+import type {NextRequest} from 'next/server';
 import {authMiddleware} from 'next-firebase-auth-edge';
 import {refreshCredentials} from 'next-firebase-auth-edge/lib/next/cookies';
 
@@ -111,7 +111,8 @@ async function refreshCredentials(
 Use `refreshNextResponseCookies` from `next-firebase-auth-edge/lib/next/cookies` to refresh authentication cookies after updating user token in [API Route Handlers](https://nextjs.org/docs/app/building-your-application/routing/route-handlers)
 
 ```tsx
-import {NextRequest, NextResponse} from 'next/server';
+import {NextResponse} from 'next/server';
+import type { NextRequest } from 'next/server';
 import {refreshNextResponseCookies} from 'next-firebase-auth-edge/lib/next/cookies';
 import {getFirebaseAuth, getTokens} from 'next-firebase-auth-edge';
 
@@ -172,7 +173,7 @@ export async function POST(request: NextRequest) {
 Use `refreshNextResponseCookiesWithToken` from `next-firebase-auth-edge/lib/next/cookies` to refresh authentication cookies based on token string extracted from Authorization header
 
 ```tsx
-import {NextRequest} from 'next/server';
+import type {NextRequest} from 'next/server';
 import {refreshNextResponseCookiesWithToken} from 'next-firebase-auth-edge/lib/next/cookies';
 
 const commonOptions = {
@@ -212,11 +213,11 @@ export async function POST(request: NextRequest) {
 
 ## Refresh auth cookies in Pages Router API Routes
 
-Use `refreshApiResponseCookies` from `next-firebase-auth-edge/lib/next/cookies` to refresh authentication cookies after updating user token in [API Routes](https://nextjs.org/docs/pages/building-your-application/routing/api-routes)
+Use `refreshApiResponseCookies` from `next-firebase-auth-edge/lib/next/api` to refresh authentication cookies after updating user token in [API Routes](https://nextjs.org/docs/pages/building-your-application/routing/api-routes)
 
 ```tsx
 import {NextApiRequest, NextApiResponse} from 'next';
-import {refreshApiResponseCookies} from 'next-firebase-auth-edge/lib/next/cookies';
+import {refreshApiResponseCookies} from 'next-firebase-auth-edge/lib/next/api';
 
 export default async function handler(
   req: NextApiRequest,
