+++ examples/next-typescript-starter/app/api/custom-claims/route.ts
@@ -1,7 +1,8 @@
 import {getFirebaseAuth} from 'next-firebase-auth-edge/lib/auth';
 import {refreshNextResponseCookies} from 'next-firebase-auth-edge/lib/next/cookies';
 import {getTokens} from 'next-firebase-auth-edge/lib/next/tokens';
-import {NextRequest, NextResponse} from 'next/server';
+import {NextResponse} from 'next/server';
+import type { NextRequest } from 'next/server';
 import {authConfig} from '../../../config/server-config';
 
 const {setCustomUserClaims, getUser} = getFirebaseAuth({
