+++ src/next/cookies.ts
@@ -1,9 +1,9 @@
-import {CookieSerializeOptions, serialize} from 'cookie';
-import {IncomingHttpHeaders} from 'http';
-import {NextApiRequest, NextApiResponse} from 'next';
-import {ReadonlyRequestCookies} from 'next/dist/server/web/spec-extension/adapters/request-cookies';
-import {RequestCookies} from 'next/dist/server/web/spec-extension/cookies';
-import {NextRequest, NextResponse} from 'next/server';
+import {type CookieSerializeOptions, serialize} from 'cookie';
+import type {IncomingHttpHeaders} from 'http';
+import type {ReadonlyRequestCookies} from 'next/dist/server/web/spec-extension/adapters/request-cookies';
+import type {RequestCookies} from 'next/dist/server/web/spec-extension/cookies';
+import {NextResponse} from 'next/server';
+import type {NextRequest} from 'next/server';
 import {getFirebaseAuth} from '../auth';
 import {SignedCookies, signCookies, signTokens} from '../auth/cookies/sign';
 import {ServiceAccount} from '../auth/credential';
@@ -77,18 +77,6 @@ export function isCookiesObjectVerifiedByMiddleware(cookies: CookiesObject) {
   );
 }
 
-export async function appendAuthCookiesApi(
-  response: NextApiResponse,
-  tokens: CustomTokens,
-  options: SetAuthCookiesOptions
-) {
-  const cookies = await signCookies(tokens, options.cookieSignatureKeys);
-
-  serializeCookies(cookies, options, (value) => {
-    response.setHeader('Set-Cookie', [value]);
-  });
-}
-
 function generateEmptyCookies(
   options: RemoveAuthCookiesOptions,
   callback: (name: string) => void
@@ -113,7 +101,7 @@ function generateCookies(
   callback(`${options.cookieName}.sig`, cookies.signature);
 }
 
-function serializeCookies(
+export function serializeCookies(
   cookies: SignedCookies,
   options: SetAuthCookiesOptions,
   callback: (setCookieHeader: string) => void
@@ -295,48 +283,6 @@ export async function verifyApiCookies(
   });
 
   return verifyTokenResult;
-}
-
-export async function refreshApiCookies(
-  cookies: Partial<{
-    [key: string]: string;
-  }>,
-  headers: IncomingHttpHeaders,
-  options: SetAuthCookiesOptions
-): Promise<VerifiedTokens> {
-  const referer = headers['referer'] ?? '';
-  const tokens = await getCookiesTokens(cookies, options);
-  const {handleTokenRefresh} = getFirebaseAuth({
-    serviceAccount: options.serviceAccount,
-    apiKey: options.apiKey,
-    tenantId: options.tenantId
-  });
-
-  const tokenRefreshResult = await handleTokenRefresh(tokens.refreshToken, {
-    referer
-  });
-
-  return {
-    customToken: tokenRefreshResult.customToken,
-    idToken: tokenRefreshResult.idToken,
-    refreshToken: tokenRefreshResult.refreshToken,
-    decodedIdToken: tokenRefreshResult.decodedIdToken
-  };
-}
-
-export async function refreshApiResponseCookies(
-  request: NextApiRequest,
-  response: NextApiResponse,
-  options: SetAuthCookiesOptions
-): Promise<NextApiResponse> {
-  const customTokens = await refreshApiCookies(
-    request.cookies,
-    request.headers,
-    options
-  );
-  await appendAuthCookiesApi(response, customTokens, options);
-
-  return response;
 }
 
 export async function verifyNextCookies(
